+++ .gitignore
@@ -5,3 +5,4 @@ run/
 .vscode/
 bin/
 *.sh
+.architectury-transformer/
