+++ common/build.gradle
@@ -1,0 +1,43 @@
+plugins {
+    id "java"
+    id "fabric-loom" version "1.10-SNAPSHOT"
+    id "com.gradleup.shadow" version "9.2.2"
+}
+
+group = project.maven_group
+version = project.mod_version
+
+java {
+    toolchain.languageVersion = JavaLanguageVersion.of(21)
+    withSourcesJar()
+}
+
+repositories {
+    maven { url "https://repo.spongepowered.org/repository/maven-public/" }
+    maven { url "https://maven.shedaniel.me/" }
+    maven { url "https://maven.terraformersmc.com/releases" }
+}
+
+loom {
+    mixin.defaultRefmapName = "common-cloud_tweaks.refmap.json"
+}
+
+remapJar {
+    dependsOn(shadowJar)
+    input.set(shadowJar.archiveFile)
+    archiveClassifier = null
+
+    from rootProject.file("LICENSE.md")
+}
+
+dependencies {
+    minecraft "com.mojang:minecraft:${rootProject.minecraft_version}"
+    implementation "org.spongepowered:mixin:0.8.5"
+    mappings loom.officialMojangMappings()
+
+    modCompileOnly "me.shedaniel.cloth:cloth-config-fabric:${project.fabric_cloth_config_version}"
+}
+
+architectury {
+    common("fabric", "forge")
+}
