+++ common/gradle.properties
@@ -1,0 +1,2 @@
+fabric_cloth_config_version=13.0.138
+fabric_loader_version=0.16.10
