+++ common/src/main/java/net/not_thefirst/story_mode_clouds/Initializer.java
@@ -1,0 +1,9 @@
+package net.not_thefirst.story_mode_clouds;
+
+import net.not_thefirst.story_mode_clouds.config.CloudsConfiguration;
+
+public class Initializer {
+    public static void initialize() {
+        CloudsConfiguration.load();
+    }
+};
