+++ common/src/main/java/net/not_thefirst/story_mode_clouds/compat/ModChecker.java
@@ -1,0 +1,5 @@
+package net.not_thefirst.story_mode_clouds.compat;
+
+public interface ModChecker {
+    boolean isLoaded(String id);
+}
