+++ common/src/main/java/net/not_thefirst/story_mode_clouds/config/ClothConfigClass.java
