+++ common/src/main/java/net/not_thefirst/story_mode_clouds/config/ModKeybinds.java
@@ -1,0 +1,17 @@
+package net.not_thefirst.story_mode_clouds.config;
+
+import com.mojang.blaze3d.platform.InputConstants;
+
+import net.minecraft.client.KeyMapping;
+import net.not_thefirst.story_mode_clouds.api.keybindings.KeyMappingHelper;
+
+public class ModKeybinds {
+    public static KeyMapping addKeybind(String id, int key, String category) {
+        return KeyMappingHelper.registerKeyBinding(new KeyMapping(id, key, category));
+    }
+
+    public static KeyMapping OPEN_CONFIG_KEYBIND = addKeybind("key.bind", InputConstants.getKey("key.keyboard.o").getValue(), "key.categories.ui");
+
+    public static void initialize() {
+    }
+}
