+++ common/src/main/java/net/not_thefirst/story_mode_clouds/renderer/mesh_builders/PuffMeshBuilder.java
