+++ common/src/main/resources/assets/cloud_tweaks/icon.png
