+++ common/src/main/resources/cloud_tweaks.mixins.json
@@ -1,0 +1,21 @@
+{
+  "required": true,
+  "minVersion": "0.8",
+  "package": "net.not_thefirst.story_mode_clouds.mixin",
+  "refmap": "common-cloud_tweaks.refmap.json",
+  "compatibilityLevel": "JAVA_17",
+  "mixins": [
+  ],
+  "client": [
+    "LevelRendererMixin",
+    "KeyMappingAccessor",
+    "OptionsMixin",
+    "ClientMixin"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  },
+  "overwrites" : {
+    "conformVisibility" : true
+  }
+}
