+++ common/src/main/resources/pack.mcmeta
@@ -1,0 +1,6 @@
+{
+    "pack": {
+        "description": "Mod resources",
+        "pack_format": 4
+    }
+}
