+++ fabric/gradle.properties
@@ -1,0 +1,4 @@
+loader_version=0.16.10
+fabric_version=0.83.0+1.20
+modmenu_version=9.0.0
+cloth_config_version=13.0.138
