+++ fabric/src/main/java/net/not_thefirst/story_mode_clouds/compat/FabricModChecker.java
@@ -1,0 +1,10 @@
+package net.not_thefirst.story_mode_clouds.compat;
+
+import net.fabricmc.loader.api.FabricLoader;
+
+public final class FabricModChecker implements ModChecker {
+    @Override
+    public boolean isLoaded(String id) {
+        return FabricLoader.getInstance().isModLoaded(id);
+    }
+}
