+++ fabric/src/main/resources/fabric.mod.json
@@ -1,0 +1,30 @@
+{
+  "schemaVersion": 1,
+  "id": "cloud_tweaks",
+  "version": "${version}",
+  "name": "Cloud Tweaks",
+  "description": "Customize your clouds!",
+  "authors": [],
+  "license": "MIT",
+  "icon": "assets/cloud_tweaks/icon.png",
+  "environment": "client",
+  "entrypoints": {
+    "client": [
+      "net.not_thefirst.story_mode_clouds.StoryModeClouds"
+    ],
+    "modmenu": [ 
+      "net.not_thefirst.story_mode_clouds.config.ModMenuIntegration"
+    ]
+  },
+  "mixins": [
+    "mixins.cloud_tweaks.json"
+  ],
+  "depends": {
+    "fabricloader": ">=0.16.10",
+    "minecraft": ">=1.20 <=1.20.1"
+  },
+  "suggests": {
+    "cloth-config": "*",
+    "modmenu": "*"
+  }
+}
