+++ forge/gradle.properties
@@ -1,0 +1,2 @@
+forge_version = 47.1.3
+loom.platform = forge
