+++ forge/src/main/java/net/not_thefirst/story_mode_clouds/compat/ForgeModChecker.java
@@ -1,0 +1,10 @@
+package net.not_thefirst.story_mode_clouds.compat;
+
+import net.minecraftforge.fml.ModList;
+
+public final class ForgeModChecker implements ModChecker {
+    @Override
+    public boolean isLoaded(String id) {
+        return ModList.get().isLoaded(id);
+    }
+}
