+++ gradle.properties
@@ -1,16 +1,7 @@
-# Done to increase the memory available to gradle.
+# 8GB machine, pardon me
 org.gradle.jvmargs = -Xmx1G
 
-# Fabric Properties
 minecraft_version = 1.20.1
-loader_version = 0.16.10
-
-# Mod Properties
-mod_version = 0.1.0
+mod_version = 0.2.0
 maven_group = net.not_thefirst
-archives_base_name = cloud_tweaks
-
-# Dependencies
-modmenu_version = 9.0.0
-cloth_config_version = 13.0.138
-fabric_version = 0.83.0+1.20
+mod_name = cloud_tweaks
