+++ settings.gradle
@@ -1,9 +1,14 @@
 pluginManagement {
     repositories {
-        maven {
-            name = 'Fabric'
-            url = 'https://maven.fabricmc.net/'
-        }
+        maven { url 'https://maven.fabricmc.net/' }
+        maven { url "https://maven.architectury.dev/" }
+        maven { url "https://files.minecraftforge.net/maven/" }
         gradlePluginPortal()
     }
 }
+
+rootProject.name = "cloud_tweaks"
+
+include("common")
+include("fabric")
+include("forge")
