--- src/client/java/net/not_thefirst/story_mode_clouds/config/ModMenuIntegration.java
@@ -1,11 +1,0 @@
-package net.not_thefirst.story_mode_clouds.config;
-
-import com.terraformersmc.modmenu.api.ConfigScreenFactory;
-import com.terraformersmc.modmenu.api.ModMenuApi;
-
-public class ModMenuIntegration implements ModMenuApi {
-    @Override
-    public ConfigScreenFactory<?> getModConfigScreenFactory() {
-        return CloudsConfiguration::createConfigScreen;
-    }
-}
