--- src/main/resources/story_mode_clouds.mixins.json
@@ -1,18 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "net.not_thefirst.story_mode_clouds.mixin",
-  "refmap": "client-cloud_tweaks+1.20.1-refmap.json",
-  "compatibilityLevel": "JAVA_21",
-  "mixins": [
-  ],
-  "client": [
-    "LevelRendererMixin",
-    "KeyMappingAccessor",
-    "OptionsMixin",
-    "ClientMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
